Move folder constants to Folder entity
